+++ packages/node/src/ethereum/api.connection.ts
@@ -73,6 +73,8 @@ export class EthereumApiConnection
       formatted_error = EthereumApiConnection.handleDisconnectionError(e);
     } else if (e.message.startsWith(`Rate Limited at endpoint`)) {
       formatted_error = EthereumApiConnection.handleRateLimitError(e);
+    } else if (e.message.includes(`Exceeded max limit of`)) {
+      formatted_error = EthereumApiConnection.handleLargeResponseError(e);
     } else {
       formatted_error = new ApiConnectionError(
         e.name,
@@ -81,6 +83,16 @@ export class EthereumApiConnection
       );
     }
     return formatted_error;
+  }
+
+  static handleLargeResponseError(e: Error): ApiConnectionError {
+    const newMessage = `Oversized RPC node response. This issue is related to the network's RPC nodes configuration, not your application. You may report it to the network's maintainers or try a different RPC node.\n\n${e.message}`;
+
+    return new ApiConnectionError(
+      'RpcInternalError',
+      newMessage,
+      ApiErrorType.Default,
+    );
   }
 
   static handleRateLimitError(e: Error): ApiConnectionError {