+++ packages/node/src/indexer/fetch.module.ts
@@ -4,7 +4,8 @@
 import { Module } from '@nestjs/common';
 import { EventEmitter2 } from '@nestjs/event-emitter';
 import {
-  BenchmarkService,
+  PoiBenchmarkService,
+  IndexingBenchmarkService,
   MmrService,
   StoreService,
   PoiService,
@@ -14,6 +15,7 @@ import {
   SmartBatchService,
   StoreCacheService,
   PgMmrCacheService,
+  MmrQueryService,
 } from '@subql/node-core';
 import { SubqueryProject } from '../configure/SubqueryProject';
 import { EthereumApiConnection } from '../ethereum/api.connection';
@@ -126,7 +128,8 @@ import { UnfinalizedBlocksService } from './unfinalizedBlocks.service';
       ],
     },
     FetchService,
-    BenchmarkService,
+    IndexingBenchmarkService,
+    PoiBenchmarkService,
     {
       provide: DictionaryService,
       useFactory: async (project: SubqueryProject, nodeConfig: NodeConfig) => {
@@ -143,6 +146,7 @@ import { UnfinalizedBlocksService } from './unfinalizedBlocks.service';
     DynamicDsService,
     PoiService,
     MmrService,
+    MmrQueryService,
     PgMmrCacheService,
     {
       useClass: ProjectService,
@@ -150,6 +154,6 @@ import { UnfinalizedBlocksService } from './unfinalizedBlocks.service';
     },
     UnfinalizedBlocksService,
   ],
-  exports: [StoreService, MmrService, StoreCacheService],
+  exports: [StoreService, MmrService, StoreCacheService, MmrQueryService],
 })
 export class FetchModule {}